Add multicart demo
The concept of having one cart communicate with another cart's piqued my interest for a while. I don't know if I'd do anything with it at this point, but I wanted to at least prove that it's possible. The latency's pretty minimal from what I've tested, so I think this has a lot of potential!